### 2022-08-22 #4698 @editorialbot
[PRE REVIEW]: AmpTorch: A Python package for scalable fingerprint-based neural network training on multi-element systems with integrated uncertainty quantification
**Submitting author:** @ajmedford (<a href="http://orcid.org/0000-0001-8311-9581">Andrew Medford</a>)
**Repository:** https://github.com/ulissigroup/amptorch
**Branch with paper.md** (empty if default branch): master
**Version:** v0.1
