### 2022-09-07 #4698 @kthyng
@ajmedford Sorry it just depends on how quickly submissions move through for the particular editors that are appropriate for your area, so it's hard to estimate.

@dhhagan Do you do any machine learning? Interested in editing this submission?
