### 2022-09-09 #4698 @dhhagan
Hey @kthyng yes - I can get to it next Tuesday if that works?
