### 2022-11-03 c1cbe36
Add nn config by turning off pin_memory for SF
