+++ amptorch/trainer.py
@@ -287,6 +287,32 @@ class AtomsTrainer:
                 verbose=self.config["cmd"].get("verbose", True),
                 **self.optimizer,
             )
+        # turn off pin memory for gaussian symmetry function force training, as torch.geometric outputs error
+        elif self.config["dataset"].get(
+            "fp_scheme", "gaussian"
+        ).lower() == "gaussian" and self.config["model"].get("get_forces", True):
+            self.net = NeuralNetRegressor(
+                module=self.model,
+                criterion=self.criterion,
+                criterion__force_coefficient=self.config["optim"].get(
+                    "force_coefficient", 0
+                ),
+                criterion__loss=self.config["optim"].get("loss", "mse"),
+                lr=self.config["optim"].get("lr", 1e-1),
+                batch_size=self.config["optim"].get("batch_size", 32),
+                max_epochs=self.config["optim"].get("epochs", 100),
+                iterator_train__collate_fn=self.parallel_collater,
+                iterator_train__shuffle=True,
+                iterator_train__pin_memory=False,
+                iterator_valid__collate_fn=self.parallel_collater,
+                iterator_valid__shuffle=False,
+                iterator_valid__pin_memory=False,
+                device=self.device,
+                train_split=self.split,
+                callbacks=self.callbacks,
+                verbose=self.config["cmd"].get("verbose", True),
+                **self.optimizer,
+            )
         else:
             self.net = NeuralNetRegressor(
                 module=self.model,
