### 2022-11-14 44986bb
Added compiler flags for PACE compiler
