+++ amptorch/descriptor/Gaussian/libsymf_builder.py
@@ -31,6 +31,7 @@ ffibuilder.set_source(
     ],
     source_extension=".cpp",
     include_dirs=["amptorch/descriptor/Gaussian/"],
+    extra_compile_args=["-no-multibyte-chars"],
 )
 
 if __name__ == "__main__":
