+++ amptorch/descriptor/GMP/libgmp_builder.py
@@ -32,6 +32,7 @@ ffibuilder.set_source(
     ],
     source_extension=".cpp",
     include_dirs=["amptorch/descriptor/GMP/"],
+    extra_compile_args=["-no-multibyte-chars"],
 )
 
 if __name__ == "__main__":
