### 2022-12-28 ulissigroup/amptorch#108 @ml-evs
Ah, just spotted that the conda env is pinning to Python 3.6, which reached end-of-life in 2021. This constraint will definitely need to be relaxed so that others can use the package.
