### 2023-01-17 fd76846
Update build to large resource class in config.yml
