+++ .circleci/config.yml
@@ -10,7 +10,8 @@ jobs:
 
       - restore_cache:
           keys:
-          - v0.2-dependencies-{{ checksum "env_cpu.yml" }}-{{ checksum "env_gpu.yml" }}
+            - v0.2-dependencies-{{ checksum "env_cpu.yml" }}-{{ checksum
+              "env_gpu.yml" }}
 
       - run:
           name: Install conda
@@ -54,3 +55,5 @@ jobs:
             conda activate amptorch
             pip install black
             black . --check
+    # The resource_class feature allows configuring CPU and RAM resources for each job. Different resource classes are available for different executors. https://circleci.com/docs/2.0/configuration-reference/#resourceclass
+    resource_class: large
