### 2023-01-17 ulissigroup/amptorch#108 @ajmedford
@ml-evs Thanks for your efforts on this, and apologies for the delay. We have run into some serious dependency version conflicts and would like your advice on how to proceed.

In short, it seems that some of the core dependencies (pytorch-geometric and pytoch-sparse) have a version conflict with Python >3.6. @nicoleyghu can elaborate more, but it seems the most practical solution is to pin the versions for everything at Python 3.6 for now and hope that we can upgrade once the dependencies become compatible with newer versions. Let us know what you think.
