### 2023-01-17 ulissigroup/amptorch#108 @ml-evs
Thanks for checking this out.

From the point of view of the JOSS, I don't think there is any hard editorial requirement that submissions do not use deprecated packages, provided the package is installable at the time of review. If you can provide a working conda environment with fully pinned dependencies on 3.6 then I can give it a go.

From my point of view, I really think it would be worth you putting in the effort to fix the dependency issues (though I really do feel your pain, believe me).  In terms of advice for how to proceed with that, you should probably test out (with fresh environments each time) the version ranges you can get away with for the major dependencies, e.g., remove all version pins in your env files (except Python), see which versions get installed, and then test them. If you already know about constraints from your own code (e.g., using a particular version of the torch API), you could include these as minimum/maximum version requirements, rather than hard pins, then try to let conda figure things out. If you want a faster dependency solver you can also try mamba (as linked in my top comment), though this won't get you around the bottleneck of downloading multiple torch versions...
