### 2023-01-17 ulissigroup/amptorch#108 @ajmedford
Thanks, I will have @nicoleyghu prioritize getting a stable environment with Python 3.6. Then she can work through some of the suggestions for improving the dependency issues as you work through the review. Fully fixing them may be beyond our skills/bandwidth if it involves modifying the pytorch-geometric or -sparse source codes, but it seems like there are some good work-arounds for at least minimizing the impact and improving the robustness of the install process.

@nicoleyghu - please let @ml-evs know when to give this another go.
