+++ env_gpu.yml
@@ -1,31 +1,28 @@
 name: amptorch
 channels:
   - pytorch
-  - conda-forge
+  - nvidia
   - defaults
 dependencies:
-  - ase=3.21*
-  - matplotlib=3.3*
-  - pip
-  - pre-commit=2.2*
-  - python=3.6.*
-  - pytorch=1.9.0
-  - tqdm=4.45.*
-  - cudatoolkit=11.1
+  - python=3.9.*
+  - pytorch=1.10.*
+  - cudatoolkit=11.3
   - h5py
-  - pykdtree
+  - matplotlib=3.3.*
+  - pip
+  - tqdm=4.45.*
   - pip:
+    - ase==3.21.*
     - cffi
-    - wandb
     - lmdb==1.0.0
+    - pre-commit
+    - pykdtree
     - skorch==0.10.*
-    - torch-geometric
-    - -f https://data.pyg.org/whl/torch-1.9.0+cu111.html
-    - torch-scatter
-    - -f https://data.pyg.org/whl/torch-1.9.0+cu111.html
-    - torch-sparse
-    - -f https://data.pyg.org/whl/torch-1.9.0+cu111.html
     - torch-cluster
-    - -f https://data.pyg.org/whl/torch-1.9.0+cu111.html
-    - torch-spline-conv
-    - -f https://data.pyg.org/whl/torch-1.9.0+cu111.html
+    - -f https://data.pyg.org/whl/torch-1.10.0+cu113.html
+    - torch-geometric
+    - -f https://data.pyg.org/whl/torch-1.10.0+cu113.html
+    - torch-scatter
+    - -f https://data.pyg.org/whl/torch-1.10.0+cu113.html
+    - torch-sparse
+    - -f https://data.pyg.org/whl/torch-1.10.0+cu113.html
