### 2023-01-24 ulissigroup/amptorch#108 @nicoleyghu
@ml-evs Thanks for your comment and suggestions! To update on the environment, the latest commit addressed deprecated Python 3.6 by upgrading to Python 3.9. Please pull again and give the updated `env_cpu.yml ` a go. I tested on OpenCI's docker image and several different machines. The average build time should be around ~13mins. Please let us know if you run into any issues.
