### 2023-01-26 ulissigroup/amptorch#108 @ml-evs
Thanks @nicoleyghu, I can confirm the package now installs fine with those instructions. I'm now just testing the GPU env too. I've been able to run most of the examples with a bit of tweaking, (e.g., editing the path to a psp file in the GMP example, and making a missing `data2.lmdb` in the LMDB examples).

I'll now start my review proper and will play around with the package over the next couple of weeks (and try to craft my own examples).

I will default to providing feedback (e.g., on the paper, docs and so on) through issues on this repo --- please let me know if you find this disruptive of the package's development and I can instead switch to writing comments on the JOSS review issue.
