+++ amptorch/trainer_config.py
@@ -1,0 +1,289 @@
+from typing import List, Dict, Union, Optional
+from dataclasses import dataclass
+import os
+import ase
+import torch
+
+
+@dataclass
+class Config_Model:
+    """
+    Configuration to define the atomistic neural network model configuration.
+
+    Attributes:
+    ------------
+        name (str): The model to be used for atomistic neural network force field, "SingleNN" or "BPNN". (default:  "SingleNN")
+
+        num_layers (int): No. of hidden layers
+
+        num_nodes (int): No. of nodes per layer
+
+        get_forces (bool): Whether to compute per-atom forces for force training. Set "force_coefficient" in "Optim" accordingly. (default: True)
+
+        batchnorm (bool): Enable batch-normalization (default:False)
+
+        activation (object): Activation function. Any activation supported by torch. (default: nn.Tanh)
+
+        **custom_args: Any additional arguments used to customize existing/new models
+    """
+
+    name: str
+    num_layers: int
+    num_nodes: int
+    get_forces: bool = True
+    batchnorm: bool = True
+    activation: object = None
+    custom_args: Dict[str, Union[int, bool, object]] = None
+
+
+@dataclass
+class Config_Optim:
+    """
+    Configuration to define the resources, and setting for neural network optimization.
+
+    Attributes:
+    ------------
+    gpus (int): No. of gpus to use, 0 for cpu (default: 0)
+
+    force_coefficient (float): If force training, coefficient to weight the force component by (default: 0)
+
+    lr (float): Initial learning rate (default: 1e-1)
+
+    batch_size (int): Batch size (default: 32)
+
+    epochs (int): Max training epochs (default: 100)
+
+    optimizer (object): Training optimizer (default: torch.optim.Adam)
+
+    loss_fn (object): Loss function to optimize (default: CustomLoss)
+
+    loss (str): Control loss function criterion, "mse" or "mae" (default: "mse")
+
+    metric (str): Metrics to be reported by, "mse" or "mae" (default: "mae")
+
+    cp_metric (str): Property based on which the model is saved. "energy" or "forces" (default: "energy")
+
+    scheduler (dict): Learning rate scheduler to use
+        {"policy": "StepLR", "params": {"step_size": 10, "gamma": 0.1}}
+    """
+
+    gpus: int = 0
+    force_coefficient: float = 0
+    lr: float = 1e-1
+    batch_size: int = 32
+    epochs: int = 100
+    optimizer: object = torch.optim.Adam
+    loss: str = "mse"
+    metric: str = "mae"
+    cp_metric: str = "energy"
+    scheduler: object = None
+
+
+@dataclass
+class Config_Dataset:
+    """
+    Configuration to define the dataset used for training, and featurization scheme to be used. Featurization default to Guassian Multipole Fingerprinting.
+
+    Attributes:
+    ------------
+    raw_data (str or List[ase.Atoms]]): Path to ASE trajectory or database, or list of Atoms objects.
+
+    lmdb_path (Optional: str): Path to LMDB database file for dataset too large to fit in memory, if raw_data is not provided.
+
+    val_split (float): Proportion of training set to use for validation.
+
+    elements (Optional: List[str]): List of unique elements in dataset, optional. Example: ["H", "O"] for water dataset.
+
+    fp_scheme (str): Fingerprinting scheme to feature dataset, "gmpordernorm" or "gaussian" (default: "gmpordernorm").
+
+    fp_params (dict): Fingerprint parameters, dataclass "GMP_params" or "SF_params".
+
+    cutoff_params (Optional, dict): Cutoff function for Gaussian fingerprinting scheme only - polynomial or cosine,
+        polynomial - {"cutoff_func": "Polynomial", "gamma": 2.0},
[…]
