+++ amptorch/descriptor/utils/pseudodensity_psp_v3/lookup_utils.py
@@ -1,0 +1,16 @@
+import os
+import json
+
+
+def generate_lookup_dictionary():
+    lookup = {}
+    for filename in os.listdir("."):
+        if filename.endswith(".g"):
+            element = filename.split("_")[0]
+            lookup[element] = filename
+
+    with open("lookup.json", "w") as f:
+        json.dump(lookup, f)
+
+
+generate_lookup_dictionary()
