+++ amptorch/trainer.py
@@ -47,11 +47,23 @@ except NameError:
 
 
 class AtomsTrainer:
-    def __init__(self, config={}):
+    """
+    Main trainer class to define the atomistic neural network force field for energy (and force prediction).
+
+    Attribute:
+    -----------------
+    config: [dict]
+        A dictionary that defines configuration of the trainer in `model`, `optim`, `dataset` and `cmd` parts. Please refer to `Usage` and `Example` sections in documentation for more information.
+    """
+
+    def __init__(self, config=None):
         self.config = config
         self.pretrained = False
 
     def load(self, load_dataset=True):
+        """
+        Loading the parameters passed from `config`.
+        """
         self.load_config()
         self.load_rng_seed()
         if load_dataset:
@@ -64,6 +76,9 @@ class AtomsTrainer:
         self.load_skorch()
 
     def load_config(self):
+        """
+        Set up attributes from input configuration dictionary.
+        """
         dtype = self.config["cmd"].get("dtype", torch.DoubleTensor)
         torch.set_default_tensor_type(dtype)
         self.timestamp = datetime.datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
@@ -89,6 +104,9 @@ class AtomsTrainer:
             os.makedirs(self.cp_dir, exist_ok=True)
 
     def load_rng_seed(self):
+        """
+        Load a fixed random seed for reproducibility.
+        """
         seed = self.config["cmd"].get("seed", 0)
         random.seed(seed)
         np.random.seed(seed)
@@ -98,6 +116,9 @@ class AtomsTrainer:
         torch.backends.cudnn.benchmark = False
 
     def get_unique_elements(self, training_images):
+        """
+        Get a list of chemical elements in str if not given in config.
+        """
         elements = np.array(
             [atom.symbol for atoms in training_images for atom in atoms]
         )
@@ -105,6 +126,9 @@ class AtomsTrainer:
         return elements
 
     def load_dataset(self):
+        """
+        Load dataset either in the format of a list of `ase.Atoms` objects or a path to those objects, or alternatively, the saved lmdb files with generated fingerprints.
+        """
         if "lmdb_path" in self.config["dataset"]:
             self.cache = self.config["dataset"].get("cache", "no")
             self.train_dataset = get_lmdb_dataset(
@@ -164,6 +188,9 @@ class AtomsTrainer:
         print("Loading dataset: {} images".format(len(self.train_dataset)))
 
     def load_model(self):
+        """
+        Load the parameters for atomistic neural network models from config.
+        """
         elements = list_symbols_to_indices(self.elements)
         model = self.config["model"].get("name", "bpnn").lower()
         if model == "bpnn":
@@ -188,6 +215,9 @@ class AtomsTrainer:
             )
 
     def load_extras(self):
+        """
+        Load extra commands such as validation splits, CV splits, debug mode and various callbacks.
+        """
         callbacks = []
         load_best_loss = train_end_load_best_loss(self.identifier)
         self.val_split = self.config["dataset"].get("val_split", 0)
@@ -238,9 +268,15 @@ class AtomsTrainer:
         self.callbacks = callbacks
 
     def load_criterion(self):
+        """
+        Load custom loss function in optimizaiton scheme from config.
+        """
         self.criterion = self.config["optim"].get("loss_fn", CustomLoss)
 
     def load_optimizer(self):
+        """
+        Load set-up parameters in optimizers from config.
+        """
         self.optimizer = {
             "optimizer": self.config["optim"].get("optimizer", torch.optim.Adam)
         }
@@ -249,6 +285,9 @@ class AtomsTrainer:
             self.optimizer.update(optimizer_args)
 
     def load_logger(self):
+        """
[…]
