+++ docs/amptorch.tests.rst
@@ -1,85 +1,0 @@
-amptorch.tests package
-======================
-
-Submodules
-----------
-
-amptorch.tests.consistency\_test module
----------------------------------------
-
-.. automodule:: amptorch.tests.consistency_test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.cp\_uncertainty\_calibration\_test module
---------------------------------------------------------
-
-.. automodule:: amptorch.tests.cp_uncertainty_calibration_test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.cutoff\_funcs\_test module
------------------------------------------
-
-.. automodule:: amptorch.tests.cutoff_funcs_test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.gaussian\_descriptor\_set\_test module
------------------------------------------------------
-
-.. automodule:: amptorch.tests.gaussian_descriptor_set_test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.pretrained\_test module
---------------------------------------
-
-.. automodule:: amptorch.tests.pretrained_test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.pretrained\_test\_lmdb module
---------------------------------------------
-
-.. automodule:: amptorch.tests.pretrained_test_lmdb
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.test\_script module
-----------------------------------
-
-.. automodule:: amptorch.tests.test_script
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.training\_test module
-------------------------------------
-
-.. automodule:: amptorch.tests.training_test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-amptorch.tests.training\_test\_gmp module
------------------------------------------
-
-.. automodule:: amptorch.tests.training_test_gmp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: amptorch.tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
