+++ docs/amptorch.uncertainty.rst
@@ -1,21 +1,0 @@
-amptorch.uncertainty package
-============================
-
-Submodules
-----------
-
-amptorch.uncertainty.utils module
----------------------------------
-
-.. automodule:: amptorch.uncertainty.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: amptorch.uncertainty
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
