+++ docs/conf.py
@@ -13,7 +13,8 @@
 import os
 import sys
 
-sys.path.insert(os.path.abspath("../"))
+# sys.path.insert(0, os.path.abspath('.'))
+sys.path.insert(0, os.path.abspath(".."))
 
 
 # -- Project information -----------------------------------------------------
