+++ .readthedocs.yaml
@@ -18,6 +18,5 @@ sphinx:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-  version: "3.9"
   install:
     - requirements: docs/requirements.txt
