+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+# docs/requirements.txt
+sphinx==3.2.1
+sphinx_rtd_theme==0.5.0
