+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 # docs/requirements.txt
 sphinx==3.2.1
 sphinx_rtd_theme==0.5.0
+Jinja2<3.1
