+++ docs/requirements.txt
@@ -2,3 +2,4 @@
 sphinx==3.2.1
 sphinx_rtd_theme==0.5.0
 Jinja2<3.1
+ase==3.21.*
