+++ docs/requirements.txt
@@ -3,3 +3,4 @@ sphinx==3.2.1
 sphinx_rtd_theme==0.5.0
 Jinja2<3.1
 ase==3.21.*
+skorch==0.10.*
