+++ docs/requirements.txt
@@ -4,4 +4,4 @@ sphinx_rtd_theme==0.5.0
 Jinja2<3.1
 ase==3.21.*
 skorch==0.10.*
-pytorch==1.10.*
+pytorch==1.0.2
