+++ docs/requirements.txt
@@ -6,5 +6,5 @@ ase==3.21.*
 skorch==0.10.*
 torch==1.10.*
 torch-geometric==2.2.0
-torch-scatter==2.1.0
-torch-sparse==0.6.16
+torch-scatter
+torch-sparse
