+++ .readthedocs.yaml
@@ -5,7 +5,7 @@ version: 2
 build:
   os: ubuntu-20.04
   tools:
-    python: "3.9"
+    python: "mambaforge-4.10"
     # You can also specify other tool versions:
     # nodejs: "16"
     # rust: "1.55"
