+++ .readthedocs.yaml
@@ -1,6 +1,11 @@
 # Required
 version: 2
 
+# Set the conda virenv
+conda:
+  environment: env_cpu.yml
+
+
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-20.04
@@ -11,11 +16,13 @@ build:
     # rust: "1.55"
     # golang: "1.17"
 
+# build package
+python:
+  version: 3.9
+  install:
+    - method: pip
+      path: .
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
-
-
-# Optionally set the conda virenv
-conda:
-  environment: env_cpu.yml
