+++ .readthedocs.yaml
@@ -11,8 +11,10 @@ build:
   os: ubuntu-20.04
   tools:
     python: "mambaforge-4.10"
-  commands:
-    - pip install -e .
+  jobs:
+    post_install:
+      - conda activate amptorch
+      - pip install -e .
     # You can also specify other tool versions:
     # nodejs: "16"
     # rust: "1.55"
