+++ .readthedocs.yaml
@@ -5,7 +5,6 @@ version: 2
 conda:
   environment: env_cpu.yml
 
-
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-20.04
@@ -13,12 +12,14 @@ build:
     python: "mambaforge-4.10"
   jobs:
     post_install:
-      - conda activate amptorch
+      - source activate amptorch
       - pip install -e .
     # You can also specify other tool versions:
     # nodejs: "16"
     # rust: "1.55"
     # golang: "1.17"
+
+
 
 
 # Build documentation in the docs/ directory with Sphinx
