+++ .readthedocs.yaml
@@ -12,7 +12,7 @@ build:
     python: "mambaforge-4.10"
   jobs:
     post_install:
-      - source activate amptorch
+      - mamba activate amptorch
       - pip install -e .
     # You can also specify other tool versions:
     # nodejs: "16"
