+++ .readthedocs.yaml
@@ -12,6 +12,7 @@ build:
     python: "mambaforge-4.10"
   jobs:
     post_install:
+      - mamba init
       - mamba activate amptorch
       - pip install -e .
     # You can also specify other tool versions:
