+++ .readthedocs.yaml
@@ -12,9 +12,7 @@ build:
     python: "mambaforge-4.10"
   jobs:
     post_install:
-      - bash
-      - conda activate amptorch
-      - pip install -e .
+      - python setup.py develop
     # You can also specify other tool versions:
     # nodejs: "16"
     # rust: "1.55"
