+++ examples/4_misc/get_fp_example.py
@@ -35,7 +35,7 @@ Gs = {
 elements = ["Cu", "C", "O"]
 dataset = AtomsDataset(
     images=images,
-    descriptor_setup=("gaussian", Gs, elements),
+    descriptor_setup=("gaussian", Gs, {"cutoff_func": "Cosine"}, elements),
     forcetraining=False,
     save_fps=True,
 )
