### 2023-03-23 #5035 @ajmedford
@ml-evs @dhhagan  Thank you for taking the time to review the code and for providing valuable feedback. We appreciate your thoughtful comments and suggestions, which have helped us to improve the quality of our work. Given the comments, @nicoleyghu has led the implementation of a number of changes, which are listed below.

DONE
1. Re-organized the example folder and provided a `readme.md` file that walks through the general structure to energy (and forces) tasks for the two main fingerprinting + neural network models, and constructing and training with lmdb files. (Comment #3)
2. Completed the documentation for main functions and classes for the module and integrated it into the documentation (see [API](https://amptorch.readthedocs.io/en/latest/amptorch.html) here). (Comment #5)
3. Made a default method for loading fitted pseudo-densities so the general users don't have to import it manually while allowing importing if specifically specified. (Comment #6)
4. Imported all test scripts into `unittest`, and made all relevant changes. The tests and examples should work now. (Comment #7)

There are also a few points that we have not completed yet, and the discussion is below:

1. [DOING] Comments #1 and #2: We are currently working on integrating the dataclass(es) for `config` of `AtomsTrainer` into the code, but as it requires refactoring how the code loads from Python Dictionary, it would probably happen for a later release.
2. Comment #2: We hope the example scripts now provide a better starting point for the users as to GMP+SingleNN or SF+BPNN's training schemes.
3. Comment #3: The key findings in GMP and Uncertainty Quantification paper require training with a larger dataset that is a bit difficult to incorporate into a demo, but the general training pertains to what is shown in the `examples` folder just with hyperparameter optimized for the given task. Hopefully this is sufficient.
4. Comment #8: We would like to finalize a release after hearing the feedback from reviewer(s).

Thank you again for your time and effort in reviewing our code and we look forward to your continued feedback. @dhhagan - please advise on the remainder of the peer review process. We would like to wait until the paper is fully "accepted" and then finalize the next release.
