### 2023-03-23 c4e39dd
Fixing documentations, tests, examples for JOSS review.  (#113)
* Add documentation and updated rtd.

* Update tests and rtds.

* Fix rtd path.

* Fix rtd path.

* Fix rtd path.

* Fix rtd path.

* Fix rtd path.

* Fix rtd path.

* Fix rtd path.

* Fix rtd requirements by adding ase.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Fix rtd requirements.

* Add conda env for rtd.

* Add conda e
[…]
