### 2023-04-17 #5035 @pmeier
In general, I'm available as well. However, I'm a little worried about

```
Language                     files          blank        comment           code
C++                             10           7091           1998          24319
Python                          48            982            474           5913
```

I'm not a C++ dev and thus cannot review anything there beyond the basics. A quick look through the repository reveals that this the C++ is mostly inside the `amptorch.descriptor.GMP` module and thus relatively "contained". In case my review is needed, maybe @ajmedford can comment if extensive C++ knowledge is required.
