### 2023-04-17 #5035 @ajmedford
@pmeier - I don't think the C++ should be too extensive. This is mostly just a performance optimization, and the technical details have been fairly well vetted at this point.
