### 2023-05-02 ulissigroup/amptorch#114 @ianfhunter
[JOSS] Confirmation of authorship
Hi, I'm starting to go through the JOSS review checklist (https://github.com/openjournals/joss-reviews/issues/5035).

First thing I notice is that I don't see the submitter's username (@ajmedford) in any commits to this repository.
There are 8 contributors, but 12 authors credited on the paper.

Also if you could confirm for me that @mshuaibii and @nicoleyghu are intended to be equal contributors (@mshuaibii appears to have an order of magnitude more contributions than @nicoleyghu - 480 commits vs 16, as well as those commits being a bit more substantial too.

non-code contributions are fine as per the JOSS authorship guidelines:
> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

but, I feel it would be remiss of me to not confirm that it is as intended.
