### 2023-05-02 ulissigroup/amptorch#114 @ajmedford
@ianfhunter good questions. I am the PI of Nicole's research group, and unfortunately have very little time to write code anymore. However, I have been involved with "active project direction" by consulting with @nicoleyghu @mshuaibii @zulissi and others. Similarly, @aapeterson and @zulissi are also PIs, but have been involved with directing the structure of the code (@aapeterson was the architect of the original AMP package). @mshuaibii wrote the original port of AMP to PyTorch, and has written a lot more code, while @nicoleyghu was more involved with the implementation and testing of the intellectually new features (UQ, GMP, etc.), and wrote the text part of the "paper", so we felt that the equal contributions was appropriate. We have also had contributions from one additional person since the review started, and we plan to add him to the author list. We have consulted with all co-authors, and I think everyone is on the same page, but I'm open to adjusting if any editors or other co-authors have issues.
