+++ env_cpu.yml
@@ -1,4 +1,4 @@
-name: amptorch
+name: amptorch_cpupy39
 channels:
   - pytorch
   - nvidia
@@ -11,6 +11,7 @@ dependencies:
   - pip
   - tqdm=4.45.*
   - cpuonly
+  - numpy=1.20.*
   - pip:
     - ase==3.21.*
     - cffi
