+++ env_cpu.yml
@@ -11,6 +11,7 @@ dependencies:
   - pip
   - tqdm=4.45.*
   - cpuonly
+  - numpy=1.20.*
   - pip:
     - ase==3.21.*
     - cffi
