### 2023-05-13 ulissigroup/amptorch#117 @ianfhunter
[JOSS] Paper feedback
Feedback from [JOSS review](https://github.com/openjournals/joss-reviews/issues/5035)

Overall, the paper is well written, I appreciate being able to understand it without too much prior knowledge. I have a few comments though to be addressed:

- line 27: "~10<sup>6</sup>+" - The claim is unclear. This should be either "~10<sup>6</sup>" OR "10<sup>6</sup>+". The training routine can scale to approximately that many points, or it can scale to more than that amount?

- line 50: You should probably cite lmdb

- line 54: the acronym UQ is not expanded.

- You make claims that AMPTorch can support higher amounts of datapoints than the base AMP and other existing codes (lines 37-38). It would be helpful to name some other examples than AMP and state their limits for comparison, to highlight the impact of this work versus the SOTA.
