### 2023-05-13 ulissigroup/amptorch#118 @ianfhunter
[JOSS] Community Guidelines
Feedback from [Joss Review](https://github.com/openjournals/joss-reviews/issues/5035)

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The community guidelines on this repository are minimal and could be improved.

I recommend you create a **CONTRIBUTING file** ([Github Help Page](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)). There are many template CONTRIBUTING files out there that you could lean upon, and it's built-in to GitHub so may get picked up by 3rd party tools, potential contributors may expect to see it there also.

**Contribution to the software**
Currently the README.md suggests only to use the technical method of "fork and pull"ing. There is no direction on what sort of contributions are welcome, how PR descriptions should be written, if tests/examples/documentation should be included with their contributions, or any other steps you would expect. These things may seem obvious, but if unstated, may not be observed.

**Report issues or problems with the software / Seek support**
Currently the README.md suggests only to file issues in the Issues tab of the repository.
There is no information on how a bug should be reported, if feature requests are accepted, etc.
You may find setting up some [issue templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) useful to ensure that some basic information is provided (e.g. you probably want users to supply a crash log if they're filing a bug)
