+++ CONTRIBUTING.md
@@ -1,0 +1,81 @@
+# Contributing to AmpTorch
+
+## Introduction
+
+Welcome! We're excited that you're interested in contributing to [AmpTorch](https://github.com/ulissigroup/amptorch/blob/master/CONTRIBUTING.md). Before you get started, please take a moment to review these guidelines to ensure a smooth and collaborative experience for everyone.
+
+## Table of Contents
+
+- [How to Contribute](#how-to-contribute)
+- [Code Style and Guidelines](#code-style-and-guidelines)
+- [Bug Reports](#bug-reports)
+- [Feature Requests](#feature-requests)
+- [Pull Requests](#pull-requests)
+- [License](#license)
+- [Community Guidelines](#community-guidelines)
+- [Contact](#contact)
+
+
+## How to Contribute
+
+1. Fork the repository hosted on Github.
+2. Create a new branch for your contribution: `git checkout -b your-branch-name`.
+3. Make your changes and commit them with clear and descriptive messages.
+    - If it involves change of codes, necessary tests should be developed and tested. 
+    - If it involves change of the API, update the documentation. 
+4. Make sure the test suite passes.  
+4. Push your changes to your forked repository.
+5. Submit a pull request (PR) to the main repository.
+
+
+## Code Style and Guidelines
+
+- Follow the existing code style and formatting conventions.
+- Write clear and concise code, with appropriate comments where necessary.
+- Maintain a consistent coding style throughout the project.
+- Ensure your changes pass all tests and do not introduce any linting errors.
+
+## Bug Reports
+
+If you encounter a bug or a problem, please help us by providing detailed information. Use the following format when reporting a bug using Github's [issues](https://github.com/ulissigroup/amptorch/issues):
+
+- Describe the issue clearly, including any error messages or logs.
+- Explain the steps to reproduce the problem.
+- Provide any additional information that might be relevant.
+
+## Feature Requests
+
+We welcome feature requests and ideas for improvement. To submit a feature request, please:
+
+- Clearly describe the feature or enhancement you'd like to see.
+- Explain why you believe it would be valuable.
+- Provide any additional context or examples that can help us understand your request better.
+
+## Pull Requests
+
+We appreciate your contributions! To submit a [pull request](https://github.com/ulissigroup/amptorch/pulls):
+
+- Make sure your changes are well-tested and do not break existing functionality.
+- Provide a clear and descriptive title for your PR.
+- Explain the purpose and scope of your changes in the description.
+- Reference any related issues or pull requests, if applicable.
+
+Reference resources for creating a PR: [How to Contribute to an Open Source Project on GitHub](https://egghead.io/series/how-to-contribute-to-an-open-source-project-on-github).
+
+## License
+When you submit code changes, the submission will be under the same [GNU General Public License](<https://www.gnu.org/licenses/>). If you are unsure, consult the license page. 
+
+## Community Guidelines
+
+We value a friendly and inclusive community. Please adhere to the following guidelines when participating:
+
+- Be respectful and considerate to others, regardless of their background or experience level.
+- Avoid offensive, discriminatory, or inappropriate language and behavior.
+- Engage in constructive discussions and provide helpful feedback.
+- Help others and contribute positively to the community.
+
+## Contact
+
+If you have any questions, concerns, or suggestions, please reach out to us via [yugehu@gatech.edu](mailto:yugehu@gatech.edu).
+
+Thank you for your interest in contributing! We look forward to your valuable contributions.
