### 2023-05-17 9e3cd5c
Adapt readme.md for contributing file.
