### 2023-05-17 9434581
Add contributing.md file and related changes to doc.  (#119)
* Fixing paper/

* Add CONTRIBUTING.md file

* Adapt readme.md for contributing file.

* Minor fix.
