+++ paper/paper.md
@@ -68,14 +68,14 @@ doi:
 
 Machine learning (ML) force fields use ML models to predict the energy (and forces) given a chemical system defined by a set of atomic coordinates. Fingerprint-based and graph-based ML force fields are two major categories of atomistic machine-learning models. Fingerprint-based models convert the atomic coordinates into fixed-length "feature vectors" that are used as inputs to regression models such as neural networks or kernel ridge regression, while graph-based systems convert the coordinates into a molecular graph which is directly input into a deep learning model. Fingerprint-based approaches tend to require less training data and be more interpretable, while graph-based models are typically able to handle more complex chemical systems and are more accurate in the limit of large datasets. Fingerprint-based models are widely used due to their conceptual similarity to traditional force-fields, but most existing packages for fingerprint-based neural network force fields are limited to systems with relatively few ($<5$) elements and relatively small ($<10^6$) datapoints.
 
-This work introduces `AmpTorch`, a Python/C++ package that leverages the Gaussian Multipole (GMP) fingerprinting scheme to build atomistic neural network models [@Lei:2021]. It provides an efficient training routine scalable to $\sim10^6+$ training points, is compatible with many-element systems ($50+$ elements), and supports statistically-rigorous uncertainty quantification (UQ) during the prediction step. The fingerprint-based structure provides relatively simple models compared to graph-based alternatives, and the structure of the fingerprinting scheme allows the package to handle systems with an arbitrary number of elements and arbitrary boundary conditions (isolated, periodic, semi-periodic).
+This work introduces `AmpTorch`, a Python/C++ package that leverages the Gaussian Multipole (GMP) fingerprinting scheme to build atomistic neural network models [@Lei:2021]. It provides an efficient training routine scalable to $\sim10^6$ training points, is compatible with many-element systems ($50+$ elements), and supports statistically-rigorous uncertainty quantification (UQ) during the prediction step. The fingerprint-based structure provides relatively simple models compared to graph-based alternatives, and the structure of the fingerprinting scheme allows the package to handle systems with an arbitrary number of elements and arbitrary boundary conditions (isolated, periodic, semi-periodic).
 
 
-`AmpTorch` is a PyTorch adaptation of the Atomistic Machine-learning Package `AMP`[@Khorshidi:2016]. It maintains the modular structure of `AMP`, with separate modules for generating fingerprints and training neural network models. `AmpTorch` supports standard symmetry function fingerprints and high-dimensional neural network potentials, but is able to scale to larger datasets and more complex chemical systems than `AMP` or other existing codes for feature-based ML potentials. `AmpTorch` outsources traditional boiler plate trainer code to `Skorch`[@Tietz:2017]. `Skorch` serves as a PyTorch wrapper that allows users to easily modify traditional neural network training strategies. `AmpTorch`'s use of `Skorch` allows users to trivially modify training strategies, model architectures, and hyperparameters.
+`AmpTorch` is a PyTorch adaptation of the Atomistic Machine-learning Package `AMP`[@Khorshidi:2016]. It maintains the modular structure of `AMP`, with separate modules for generating fingerprints and training neural network models. `AmpTorch` supports standard symmetry function fingerprints and high-dimensional neural network potentials, but is able to scale to larger datasets and more complex chemical systems than `AMP` or other existing codes for feature-based ML potentials (for example, Atom-Centered Symmetry Functions[@Behler:2015] whose fingerprinting dimension scales with the number of elements). `AmpTorch` outsources traditional boiler plate trainer code to `Skorch`[@Tietz:2017]. `Skorch` serves as a PyTorch wrapper that allows users to easily modify traditional neural network training strategies. `AmpTorch`'s use of `Skorch` allows users to trivially modify training strategies, model architectures, and hyperparameters.
 
[…]
