+++ paper/paper.bib
@@ -65,6 +65,25 @@
     month = {7}
 }
 
+@misc{Chu:2010,
+    title = {{Lightning Memory-Mapped Database Manager (LMDB)}},
+    year = {2010},
+    author = {'Chu, Howard', 'Hedenfalk, Martin'},
+}
+
+
+@article{Behler:2015,
+    title = {{Constructing High-Dimensional Neural Network Potentials:A Tutorial Review}},
+    year = {2015},
+    journal = {International Journal of Quantum Chemistry},
+    author = {J{\O}rg Behler},
+    pages = {1032–1050},
+    volume = {115},
+    publisher = {IOP Publishing},
+    doi = {110.1002/qua.24890},
+}
+
+
 @article{Larsen:2017,
     title = {{The atomic simulation environment—a Python library for working with atoms}},
     year = {2017},
