### 2023-05-19 ulissigroup/amptorch#112 @ajmedford
@ml-evs - I think we addressed these comments in https://github.com/openjournals/joss-reviews/issues/5035 by mistake. Let us know if any of these issues still need to be addressed before publication.
