### 2023-05-19 ulissigroup/amptorch#117 @ml-evs
Just so you have it all in once place, I'll also add my comments on the paper here. The paper is well-written and the package is well-motivated, so my suggestions are mostly minor. (Feel free to tick the boxes yourself)

- [ ] 38: boiler plate -> boilerplate
- [ ] 47: SingleN -> SingleNN
- [ ] 50: Btree-based -> B-tree-based
- [ ] 54: I think the UQ section should cite your own paper on UQ (https://arxiv.org/abs/2208.08337), at the very least!
- [ ] 58: statistically-rigorous -> statistically rigorous
- [ ] 84: the URL reference is mangled, presumably should be just https://doi.org/10.1021/acscatal.0c04525
- [ ] I would suggest adding URLs for the LMDB and Skorch references
- [ ] Minor stylistic thing but it seems like most other JOSS papers include a space before the reference, e.g., `blah blah [@Evans2023]` rather than `blah blah[@Evans2023]`

- [ ] This sentence irks me slightly:
    > Thus, AMPTorch is currently the only feature-based ML force field code capable of training on an arbitrary number of elements and training points

   as unfortunately by the time of publication these claims are almost always out-of-date (I take at least some of the blame for that!) I would suggest rephrasing it to something like "AMPTorch fills a gap in the ecosystem for feature-based ML force fields with capability for handling arbitrary numbers of elements and training points". I'm aware this is bordering on personal preference (maybe its just the italics...) though so I'll leave it up to you.
