### 2023-05-19 ulissigroup/amptorch#112 @ml-evs
Hi @ajmedford, thanks for pointing this out, and thanks to you and @nicoleyghu for the work implementing my suggestions. I've ticked off the things that are fixed above, and although I think some basic examples would be helpful I'm satisfied that the examples cover your advanced features well.

I've just been going through my initial testing scripts and I've run into a couple of minor issues.

The default `fp_scheme` seems to have changed since my last attempts, so the docs are now out of date [here](https://github.com/ulissigroup/amptorch/blob/943458177c66c11683f0a63b3c67d51d7cc9d0ea/README.md?plain=1#L60):

```diff
-     "fp_scheme": str,             # Fingerprinting scheme to feature dataset, "gaussian" or "gmp" (default: "gaussian")
+     "fp_scheme": str,             # Fingerprinting scheme to feature dataset, "gaussian", "gmp" or "gmpordernorm" (default: "gmpordernorm")
```

My old training script didn't set `fp_scheme` and now fails with the error

```python
  File "/home/mevans/src/amptorch/amptorch/dataset.py", line 49, in __init__
    self.descriptor = construct_descriptor(descriptor_setup)
  File "/home/mevans/src/amptorch/amptorch/dataset.py", line 131, in construct_descriptor
    descriptor = GMPOrderNorm(MCSHs=fp_params, elements=elements)
  File "/home/mevans/src/amptorch/amptorch/descriptor/GMPOrderNorm/__init__.py", line 35, in __init__
    self.default_cutoff()
  File "/home/mevans/src/amptorch/amptorch/descriptor/GMPOrderNorm/__init__.py", line 59, in default_cutoff
    sigmas = self.MCSHs["MCSHs"]["sigmas"]
```

There is of course no requirement that your API stays fixed during review, but just thought I'd mention it in case it was unintentional! Setting `fp_scheme` to `gaussian` allows me to train potentials once again.
