### 2023-05-19 bb639af
Fix readme.md and check the default values with AtomsTrainer class.
