### 2023-05-22 2877ec5
Fix issues in paper.md per Issue #117
