### 2023-06-04 a1ce16b
Updated usage.rst for rtd
