### 2023-06-05 f543162
Update readthedoc documentation (#121)
* Fixing paper/

* Add CONTRIBUTING.md file

* Adapt readme.md for contributing file.

* Minor fix.

* Fix readme.md and check the default values with AtomsTrainer class.

* Fix issues in paper.md per Issue #117

* Usage documentation update.

* Fix format via black.

* Updated usage.rst for rtd

* Add image to docs/

* Add to 2D water example.
