+++ docs/usage.rst
@@ -4,6 +4,144 @@
 Usage
 ==================================
 
+Training and Predicting with 2D Water Trajectory
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+In this section, we demonstrate how to train a NNFF with AmpTorch on an ase 
+trajectory of a water molecule and visualize the potential energy surface (PES). 
+In the trajectory of a single water molecule, one O-H bond length is fixed, and the 
+other O-H length and the H-O-H bond angle are allowed to vary, hence 2D PES. 
+This dataset is calculated by Quantem Espresso (xc: PBE. kpts: (1,1,1). Planewave: 500). 
+
+To start, we load all images in the trajectory: 
+
+.. code-block:: python
+
+   import matplotlib.pyplot as plt
+   import ase
+   import ase.io
+   from amptorch.ase_utils import AMPtorch
+   from amptorch.trainer import AtomsTrainer
+
+   # read all images from the trajectory
+   images = ase.io.read("./water_2d.traj", index=":") # no train-test/holdout split for demo purpose
+
+We skipped train-test-split step here for the simplicity of a demo case. 
+
+Then, we define the hyperparameters of GMP fingerprinting scheme in radial probes(nsigmas) and their range (max_radial_sigma), and angular probes (max_MCSH_order): 
+
+.. code-block:: python
+
+   sigmas = np.linspace(0, max_radial_sigma, nsigmas + 1, endpoint=True)[1:]
+   GMPs = {
+      "MCSHs": {"orders": list(range(max_MCSH_order + 1)), "sigmas": sigmas},
+   }
+
+Next, we define the training configuration: 
+
+.. code-block:: python
+
+   config = {
+      "model": {
+         "name": "singlenn",
+         "get_forces": True,
+         # "num_layers": 3,
+         # "num_nodes": 20,
+         "hidden_layers": [20,20,20], # more flexible way of defining NN, alternative to define both "num_layers" and "num_nodes"
+         "activation": torch.nn.Tanh,
+         "batchnorm": True,
+      },
+      "optim": {
+         "device": "cpu",
+         "force_coefficient": 0.01,
+         "lr": 1e-3,
+         "batch_size": 16,
+         "epochs": 500,
+         "loss": "mse",
+         "metric": "mae",
+      },
+      "dataset": {
+         "raw_data": images,
+         "val_split": 0,
+         "fp_scheme": "gmpordernorm",
+         "fp_params": GMPs,
+         "save_fps": False,
+      },
+      "cmd": {
+         "debug": False,
+         "run_dir": "./",
+         "seed": 1,
+         "identifier": "2d_water",
+         "verbose": True,
+         "logger": False,
+      },
+   }
+
+To start training: 
+
+.. code-block:: python
+
+   # train
+   torch.set_num_threads(1)
+   trainer = AtomsTrainer(config)
+   trainer.train()
+
+Finally, we'd like to make predictions on how the potential energy changes with varying O-H bond length with the H-O-H angle at around equilibrium bond angle (104.21 degrees). We first obtained the ground truth DFT values from loaded trajectory, and compared with NNFF predictions. 
+
+.. code-block:: python
+
+   # ground truth DFT 
+   training_angle100 = [_ for _ in images if np.isclose(_.get_angle(1, 0, 2), 104.210, atol=1e-3)]
+
+   distances_training = [_.get_distance(0, 2) for _ in training_angle100]
+   energies_training = [_.get_potential_energy() for _ in training_angle100]
+
+   # predicted NNFF
+   distances = np.linspace(0.4, 2.0, 100)
+   images = []
+   for dist in distances:
[…]
