### 2023-06-05 326c6b1
Minor fixes to paper and docs.
