+++ paper/paper.md
@@ -83,7 +83,7 @@ In addition, `AmpTorch` leverages the B-tree-based database management library, 
 
 # Statement of need
 
-There are numerous software packages for both constructing ML force-fields. These include packages that are based on deep learning and graph convolutional models, kernel-based regression models, and neural networks with fixed feature vectors. Of these packages, only graph convolutional codes have been demonstrated for extremely large and complex datasets with $>$ 10 elements and $>$ 1M training points such as [Open Catalyst 2020 Dataset](https://pubs.acs.org/doi/10.1021/acscatal.0c04525) [@Chanussot:2021]. However, at inference time, adapting these often complex architectures with 1-100M parameter models into atomistic simulation interfaces seeking extremely fast calls is a challenging task. On the other hand, all existing packages for training feature-based models are limited in both the number of elements they can handle (due to the poor scaling of feature vector size) and the number of training points (due to scaling of training kernel-based models and memory management issues in most codes). Thus, `AMPTorch` is _currently the only feature-based ML force field code capable of training on an arbitrary number of elements and training points._ For this reason, we expect that the code will be widely used by researchers seeking to train ML force-field models for complex systems that can be integrated with existing atomistic simulation pipelines.
+There are numerous software packages for both constructing ML force-fields. These include packages that are based on deep learning and graph convolutional models, kernel-based regression models, and neural networks with fixed feature vectors. Of these packages, only graph convolutional codes have been demonstrated for extremely large and complex datasets with $>$ 10 elements and $>$ 1M training points such as [Open Catalyst 2020 Dataset](https://pubs.acs.org/doi/10.1021/acscatal.0c04525) [@Chanussot:2021]. However, at inference time, adapting these often complex architectures with 1-100M parameter models into atomistic simulation interfaces seeking extremely fast calls is a challenging task. On the other hand, all existing packages for training feature-based models are limited in both the number of elements they can handle (due to the poor scaling of feature vector size) and the number of training points (due to scaling of training kernel-based models and memory management issues in most codes). Thus, `AMPTorch` aims to fill in gap of developing a toolkit to train fingerprint-based neural network force fields on datasets of an arbitrary number of chemical elements. For this reason, we expect that the code will be widely used by researchers seeking to train ML force-field models for complex systems that can be integrated with existing atomistic simulation pipelines.
 
 # Acknowledgements
 
