+++ docs/usage.rst
@@ -4,6 +4,151 @@
 Usage
 ==================================
 
+Training and Predicting with 2D Water Trajectory
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+In this section, we demonstrate how to train a NNFF with AmpTorch on an ase 
+trajectory of a water molecule and visualize the potential energy surface (PES). 
+In the trajectory of a single water molecule, one O-H bond length is fixed, and the 
+other O-H length and the H-O-H bond angle are allowed to vary, hence 2D PES. 
+This dataset is calculated by Quantem Espresso (xc: PBE. kpts: (1,1,1). Planewave: 500). 
+
+To start, we load all images in the trajectory: 
+
+.. code-block:: python
+
+   import matplotlib.pyplot as plt
+   import ase
+   import ase.io
+   from amptorch.ase_utils import AMPtorch
+   from amptorch.trainer import AtomsTrainer
+
+   # read all images from the trajectory
+   images = ase.io.read("./water_2d.traj", index=":") # no train-test/holdout split for demo purpose
+
+We skipped train-test-split step here for the simplicity of a demo case. 
+
+Then, we define the hyperparameters of GMP fingerprinting scheme in radial probes(nsigmas) and their range (max_radial_sigma), and angular probes (max_MCSH_order): 
+
+.. code-block:: python
+
+   sigmas = np.linspace(0, max_radial_sigma, nsigmas + 1, endpoint=True)[1:]
+   GMPs = {
+      "MCSHs": {"orders": list(range(max_MCSH_order + 1)), "sigmas": sigmas},
+   }
+
+Next, we define the training configuration with aspects including choice of model, optimizer and fingerprinting scheme: 
+
+.. code-block:: python
+
+   config = {
+      # model choices of neural network force fields
+      "model": {
+         "name": "singlenn",  # neural network architecture
+         "get_forces": True, # for structure to energy and forces (S2EF) task
+         # "num_layers": 3,
+         # "num_nodes": 20,
+         "hidden_layers": [20,20,20], # setting the dimensions of the 2nd to 4th hidden layers to 20
+         "activation": torch.nn.Tanh,
+         "batchnorm": True,
+      },
+      # optimization schemes associated with training
+      "optim": {
+         "device": "cpu", # device
+         "force_coefficient": 0.01, # adjust the loss of energy vs. forces
+         "lr": 1e-3, # learning rate
+         "batch_size": 16, # batch size in training
+         "epochs": 500, # total number of iterations of optimization
+         "loss": "mse", # loss function evaluation
+         "metric": "mae", # evaluation metrics
+      },
+      # dataset processing and fingerprinting
+      "dataset": {
+         "raw_data": images, # list of ase.Atoms objects for training (and validation)
+         "val_split": 0, # validation fraction
+         "fp_scheme": "gmpordernorm", # fingerprinting scheme
+         "fp_params": GMPs, # fingerprinting parameters
+         "save_fps": False, # whether to save fingerprints
+      },
+      # other commands
+      "cmd": {
+         "debug": False,
+         "run_dir": "./",
+         "seed": 1,
+         "identifier": "2d_water",
+         "verbose": True,
+         "logger": False,
+      },
+   }
+
+To start training: 
+
+.. code-block:: python
+
+   # train
+   torch.set_num_threads(1)
+   trainer = AtomsTrainer(config)
+   trainer.train()
+
[…]
