### 2023-06-06 #5035 @ajmedford
Sure, we can check the DOI's. A few papers have now appeared in journals so we can update those.

Not sure if this matters, but we did add one author (Jacob Paras) in between the initial submission and the final version. Perhaps this is causing an affiliation issue? Let us know if we need to make any changes.
