### 2023-06-08 fe5dd8a
[paper] Fix doi's and some citations
