+++ paper/paper.bib
@@ -1,29 +1,38 @@
-@article{Lei:2021,
-    title = {{A Universal Framework for Featurization of Atomistic Systems}},
-    year = {2021},
+@article{Lei:2022,
+    archivePrefix = {arXiv},
+    arxivId = {2102.02390},
     author = {Lei, Xiangyun and Medford, Andrew J.},
-    url = {http://arxiv.org/abs/2102.02390},
-    arxivId = {2102.02390}
+    doi = {10.1021/acs.jpclett.2c02100},
+    eprint = {2102.02390},
+    issn = {19487185},
+    journal = {Journal of Physical Chemistry Letters},
+    number = {34},
+    pages = {7911--7919},
+    pmid = {35980312},
+    title = {{A Universal Framework for Featurization of Atomistic Systems}},
+    volume = {13},
+    year = {2022}
 }
 
-@article{Hu2022,
-abstract = {Uncertainty quantification (UQ) is important to machine learning (ML) force fields to assess the level of confidence during prediction, as ML models are not inherently physical and can therefore yield catastrophically incorrect predictions. Established a-posteriori UQ methods, including ensemble methods, the dropout method, the delta method, and various heuristic distance metrics, have limitations such as being computationally challenging for large models due to model re-training. In addition, the uncertainty estimates are often not rigorously calibrated. In this work, we propose combining the distribution-free UQ method, known as conformal prediction (CP), with the distances in the neural network's latent space to estimate the uncertainty of energies predicted by neural network force fields. We evaluate this method (CP+latent) along with other UQ methods on two essential aspects, calibration, and sharpness, and find this method to be both calibrated and sharp under the assumption of independent and identically-distributed (i.i.d.) data. We show that the method is relatively insensitive to hyperparameters selected, and test the limitations of the method when the i.i.d. assumption is violated. Finally, we demonstrate that this method can be readily applied to trained neural network force fields with traditional and graph neural network architectures to obtain estimates of uncertainty with low computational costs on a training dataset of 1 million images to showcase its scalability and portability. Incorporating the CP method with latent distances offers a calibrated, sharp and efficient strategy to estimate the uncertainty of neural network force fields. In addition, the CP approach can also function as a promising strategy for calibrating uncertainty estimated by other approaches.},
-archivePrefix = {arXiv},
-arxivId = {2208.08337},
-author = {Hu, Yuge and Musielewicz, Joseph and Ulissi, Zachary W. and Medford, Andrew J.},
-doi = {10.1088/2632-2153/ACA7B1},
-eprint = {2208.08337},
-issn = {2632-2153},
-journal = {Machine Learning: Science and Technology},
-keywords = {conformal prediction,machine learning force fields,neural network force fields,uncertainty quantification},
-month = {dec},
-number = {4},
-pages = {045028},
-publisher = {IOP Publishing},
-title = {{Robust and scalable uncertainty estimation with conformal prediction for machine-learned interatomic potentials}},
-url = {https://iopscience.iop.org/article/10.1088/2632-2153/aca7b1 https://iopscience.iop.org/article/10.1088/2632-2153/aca7b1/meta},
-volume = {3},
-year = {2022}
+
+@article{Hu:2022,
[…]
