+++ paper/paper.bib
@@ -65,12 +65,18 @@
 }
 
 @article{Musielewicz:2022,
-    title = {{FINETUNA: Fine-tuning Accelerated Molecular Simulations}},
-    year = {2022},
-    author = {Musielewicz, Joseph and Wang, Xiaoxiao and Tian, Tian and Ulissi, Zachary},
-    url = {http://arxiv.org/abs/2205.01223},
+    archivePrefix = {arXiv},
     arxivId = {2205.01223},
-    keywords = {active learning, dft, fine tuning, graph potential energy surface}
+    author = {Musielewicz, Joseph and Wang, Xiaoxiao and Tian, Tian and Ulissi, Zachary},
+    doi = {10.1088/2632-2153/ac8fe0},
+    eprint = {2205.01223},
+    issn = {26322153},
+    journal = {Machine Learning: Science and Technology},
+    keywords = {DFT,active learning,fine tuning,graph potential energy surface},
+    number = {3},
+    title = {{FINETUNA: fine-tuning accelerated molecular simulations}},
+    volume = {3},
+    year = {2022}
 }
 
 @article{Liu:2020,
