### 2023-06-08 a370b71
Paper fix (#124)
* Fix doi's in paper.bib
* Fix citations in paper.md
