### 2023-06-13 #5035 @ml-evs
Just trying to help debug this @ajmedford, you can see the PDF generation errors here: https://github.com/openjournals/joss-papers/actions/runs/5253611987/jobs/9491155775

Its a bit hidden, but it looks like the remaining problem is just a typo in the affiliation setting:

```ruby
/home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.2.0/gems/theoj-1.5.4/lib/theoj/paper.rb:158:in `failure':
Author (Jacob Paras) is missing affiliation (Theoj::Error)
```

I can make a quick PR that fixes this for you, then you should be able to run the same command I did above to check for any other errors.

@dhhagan could you then advise of the next step to get this out of the door? I also note that one of the reviewers listed above (professoralkmin) did not end up reviewing the paper.
