### 2023-06-13 988ffa7
Fix affiliation typo that is breaking JOSS build
