### 2023-06-16 #5035 @ajmedford
Thanks @dhhagan! We just got feedback from the co-authors, and have the following list of minor changes that @nicoleyghu will work on getting implemented and pushed ASAP. Other than that it looks great!

L26: "relatively small datapoints" -> "relatively small datasets"
L35, 36, 39, etc.: AMP -> Amp
L70: "both constructing..." -> "both constructing and applying"
L74: "Open Catalyst 2020 Dataset" -> "datasets in the Open Catalyst Project"
L80: "fill in gap of" -> "fill this gap by"
