### 2023-06-18 b50f091
Update paper.md and reference formatting
