### 2023-06-18 30b0cdd
Update paper.md and reference formatting (#126)
