+++ paper/paper.bib
@@ -96,14 +96,14 @@
 @misc{Tietz:2017,
     title = {{skorch: A scikit-learn compatible neural network library that wraps PyTorch}},
     year = {2017},
-    author = {'Tietz, Marian', 'Fan, Thomas J.', 'Nouri, Daniel', 'Bossan, Benjamin'},
+    author = {Tietz, Marian and Fan, Thomas J. and Nouri, Daniel and Bossan, Benjamin},
     month = {7}
 }
 
 @misc{Chu:2010,
     title = {{Lightning Memory-Mapped Database Manager (LMDB)}},
     year = {2010},
-    author = {'Chu, Howard', 'Hedenfalk, Martin'},
+    author = {Chu, Howard and Hedenfalk, Martin},
 }
 
 
