### 2023-07-01 f525024
Update DOI badge placement
