+++ env_cpu.yml
@@ -11,7 +11,7 @@ dependencies:
   - pip
   - tqdm=4.45.*
   - cpuonly
-  - numpy=1.20.*
+  - numpy=1.21.2
   - pip:
     - ase==3.21.*
     - cffi
