+++ env_gpu.yml
@@ -11,7 +11,7 @@ dependencies:
   - matplotlib=3.3.*
   - pip
   - tqdm=4.45.*
-  - numpy=1.20.*
+  - numpy=1.21.2
   - pip:
     - ase==3.21.*
     - cffi
