+++ env_cpu.yml
@@ -11,7 +11,6 @@ dependencies:
   - pip
   - tqdm=4.45.*
   - cpuonly
-  - numpy=1.21.2
   - pip:
     - ase==3.21.*
     - cffi
@@ -27,3 +26,5 @@ dependencies:
     - -f https://data.pyg.org/whl/torch-1.10.0+cpu.html
     - torch-sparse==0.6.13
     - -f https://data.pyg.org/whl/torch-1.10.0+cpu.html
+    - numpy==1.20.*
+
