### 2023-07-02 0c2e0ed
Add DOI badge to readme (#127)
* Add zenodo doi badge

* Update environment

* Add zenodo affliation metadata.
