### 2023-07-03 #5035 @ajmedford
@dhhagan - We have released and tagged the version and archived it. The release is here: https://github.com/ulissigroup/amptorch/releases/tag/v0.2.0-alpha,  and the archived version is here: https://zenodo.org/record/8103323.

Let us know if you need anything else to get this finalized!
