### 2023-07-14 #5035 @dhhagan
Hi @ajmedford - sorry for the delay! Can you update the title in Zenodo to match the paper title? Also, are you planning to update the release version to no longer indicate a pre-release? If so, now would be a good time to do that so that the paper and release in GitHub/zenodo all align?
