### 2023-07-16 02e5af3
Minor fix for Zenodo (#128)
* Minor fix

* Minor fix for Zenodo update

* Delete zenodo meta file for later edit
